renames mod/dir to split_file
