+++ accounts-db/src/lib.rs
@@ -34,7 +34,7 @@ pub mod read_only_accounts_cache;
 #[cfg(not(feature = "dev-context-only-utils"))]
 mod read_only_accounts_cache;
 mod rolling_bit_field;
-mod split_accounts_file;
+mod split_file;
 pub mod stake_rewards;
 pub mod storable_accounts;
 pub mod utils;
